Bug Fixes, Plot Title and Status Indication
Fixed a shapefle creation bug where shapefiles would be created without
geospatial information. Also, added the ability to change the plot
title. This is also reflected in shapefile saves, as a shapefile with
multiple spectra in it will take the name of the plot. Finally, added
status indicator icons for indicated when a cube is attached and when
geodata is linked to the GUI.